Fix database query (again), more strucutred code, fix encoding
The `revision_actor_temp` table was dropped in production in May 2022
(T307906), so presumably this code hasn’t worked since then...

Move the query to a function, which is run just before the result is
needed. This makes the `unset()` calls unnecessary (the variables that
used to be unset are now local ones, so they’re automatically destroyed
at the end of the function) and causes at least the header to be shown
in case of an exception – the header contains the link to this repo, so
interested people have a chance to submit a PR if they figure out what
the problem is.

Remove `utf8_encode()` calls – everything should be UTF-8 by now, so
interpreting user names as ISO-8859-1 (while they are actually UTF-8)
and re-encoding them as UTF-8 breaks things rather than fixing them.